requirements.txt
Django==5.1.5
django-ninja==1.3.0
django-ninja-extra==0.22.3
django-ninja-jwt==5.3.5
psycopg2-binary==2.9.10
